moved ct_utils into md_utils folder
