+++ api/batch_processing/postprocessing/repeat_detection_elimination/find_repeat_detections.py
@@ -26,7 +26,7 @@ import argparse
 import os
 import sys
 
-import ct_utils
+from md_utils import ct_utils
 from api.batch_processing.postprocessing.repeat_detection_elimination import repeat_detections_core
