+++ api/synchronous/api_core/animal_detection_api/.gitignore
@@ -2,4 +2,5 @@ data_management/annotations/
 md_visualization/
 run_tf_detector.py
 ct_utils.py
+md_utils/
 allowed_keys.txt
