+++ detection/run_inference_with_yolov5_val.py
@@ -313,7 +313,7 @@ def run_inference_with_yolo_val(options):
 #%% Command-line driver
 
 import argparse,sys
-from ct_utils import args_to_object
+from md_utils.ct_utils import args_to_object
 
 def main():
