+++ detection/tf_detector.py
@@ -10,7 +10,7 @@
 import numpy as np
 
 from detection.run_detector import CONF_DIGITS, COORD_DIGITS, FAILURE_INFER
-from ct_utils import truncate_float
+from md_utils.ct_utils import truncate_float
 
 import tensorflow.compat.v1 as tf
