+++ megadetector.md
@@ -784,9 +784,8 @@ mkdir c:\git
 cd c:\git
 git clone https://github.com/agentmorris/MegaDetector
 cd c:\git\MegaDetector
-mamba create -n cameratraps-detector
+mamba create -n cameratraps-detector pip -y
 mamba activate cameratraps-detector
-mamba install -y pip
 pip install -r envs\requirements.txt
 set PYTHONPATH=c:\git\MegaDetector
 ```
@@ -798,9 +797,8 @@ mkdir ~/git
 cd ~git
 git clone https://github.com/agentmorris/MegaDetector
 cd ~/git/MegaDetector
-mamba create -n cameratraps-detector
+mamba create -n cameratraps-detector pip -y
 mamba activate cameratraps-detector
-mamba install -y pip
 pip install -r envs/requirements.txt
 export PYTHONPATH="$HOME/git/MegaDetector"
 ```
@@ -810,7 +808,7 @@ YMMV.
 If you're feeling even more experimental:
 
 ```batch
-pip install megadetector
+pip install megadetector --upgrade
 ```
 
 ...also works, with the same caveats as above: this will not produce results that are literally identical to the training environment, so, YMMV.  If you use this route, make sure the MegaDetector and YOLOv5 folders are <i>not</i> on your PYTHONPATH.
